Move dashboard from /dashboard-new to /dashboard route
- Replace /dashboard with the modern dashboard from /dashboard-new
- Update all navigation links and redirects to use /dashboard
- Update authentication flows to redirect to /dashboard after login
- Remove /dashboard-new directory and all references
- Maintain route protection and existing functionality
