+++ openmemory/ui/app/dashboard/layout.tsx
@@ -1,6 +1,6 @@
 import { ScrollArea } from "@/components/ui/scroll-area";
 
-export default function DashboardLayout({
+export default function DashboardNewLayout({
   children,
 }: {
   children: React.ReactNode;
@@ -10,4 +10,4 @@ export default function DashboardLayout({
       {children}
     </ScrollArea>
   );
-}
+}
